+++ app/src/main/java/sv/edu/ues/fia/pupuseria/ProductoEliminarActivity.java
@@ -23,12 +23,21 @@ public class ProductoEliminarActivity extends Activity {
     }
 
     public void eliminarCamposProducto(View v){
-        String regDel;
-        Producto p = new Producto();
-        p.setId_producto(Integer.parseInt(editIdProdDel.getText().toString()));
-        helper.abrir();
-        regDel=helper.eliminarProducto(p);
-        helper.cerrar();
-        Toast.makeText(this, regDel, Toast.LENGTH_SHORT).show();
+        try{
+            if(editIdProdDel.getText().toString().isEmpty()){
+                Toast.makeText(this, "El Campo ID Producto se esta enviando vacio, por favor completar", Toast.LENGTH_SHORT).show();
+            }else{
+                String regDel;
+                Producto p = new Producto();
+                p.setId_producto(Integer.parseInt(editIdProdDel.getText().toString()));
+                helper.abrir();
+                regDel=helper.eliminarProducto(p);
+                helper.cerrar();
+                Toast.makeText(this, regDel, Toast.LENGTH_SHORT).show();
+            }
+        }catch (Exception e){
+            Toast.makeText(this, "A ocurrido un error durante la ejecucion en Eliminar Producto", Toast.LENGTH_SHORT).show();
+        }
+
     }
 }
